revert back to function params for mandatory values of store builders
