+++ app/src/main/kotlin/pro/respawn/flowmvi/sample/provider/LambdaCounterProvider.kt
@@ -4,12 +4,13 @@ import androidx.lifecycle.SavedStateHandle
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.onEach
 import pro.respawn.flowmvi.android.plugins.androidLoggingPlugin
-import pro.respawn.flowmvi.android.plugins.saveState
+import pro.respawn.flowmvi.android.plugins.parcelizeState
 import pro.respawn.flowmvi.api.PipelineContext
 import pro.respawn.flowmvi.dsl.LambdaIntent
 import pro.respawn.flowmvi.dsl.reduceLambdas
 import pro.respawn.flowmvi.dsl.store
 import pro.respawn.flowmvi.plugins.whileSubscribed
+import pro.respawn.flowmvi.sample.provider.CounterState.Loading
 import pro.respawn.flowmvi.sample.repo.CounterRepo
 
 private typealias Context = PipelineContext<CounterState, LambdaIntent<CounterState, CounterAction>, CounterAction>
@@ -19,17 +20,16 @@ class LambdaCounterProvider(
     private val repo: CounterRepo,
 ) {
 
-    val store = store<CounterState, LambdaIntent<CounterState, CounterAction>, CounterAction> {
+    val store = store<CounterState, LambdaIntent<CounterState, CounterAction>, CounterAction>(Loading) {
         name = "Counter"
         install(androidLoggingPlugin())
         reduceLambdas()
-        saveState(savedStateHandle)
+        parcelizeState(savedStateHandle)
         whileSubscribed {
             repo.getTimer()
                 .onEach { produceState(it) } // set mapped states
                 .consume(Dispatchers.Default)
         }
-        initial(CounterState.Loading)
     }
 
     private suspend fun Context.produceState(timer: Int) = updateState {
